Modify the api's based on review comments
